requirements.txt
firecrawl-py
openai
pandas
requests
beautifulsoup4
python-dotenv
pyinstaller
tabulate
reportlab
pywin32
